Improve testing add run commands
run script works with subcommands,
gives shorthand for running the assembler
and compiler. Add functions for running the
tests in a subshell and save up the stdout
and stderr, gives the the . or the x for the
passes and fails for the tests, then once run
prints all of the stdout and stderr for each of
the tests as well as a summary of passes and fails.